DV1.0.1 : Improved Worker Management Ui and Logic

+++ app/src/main/java/com/rootlink/mystoremanager/data/dao/StockDao.kt
@@ -19,7 +19,7 @@ interface StockDao {
     """)
     suspend fun updateStock(
         productId: Long,
-        delta: Double,
+        delta: Int,
         time: Long
     )
